+++ src/assets/other/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
-- version 3
-- version 2
-- version 1
+This is the **third** iteration of my portfolio website.
+
+Codebases:
+
+- [latest](https://github.com/mvoemel/portfolio)
+- [version 2](https://github.com/mvoemel/portfolio/tree/v2)
+- [version 1](https://github.com/mvoemel/portfolio/tree/v1)